svelte: extract regressor logic from `TimersSetHistoryMetric.svelte`
